requirements.txt
streamlit>=1.40.0
google-generativeai>=0.8.0
plotly>=5.24.0
python-dotenv>=1.0.0
